pyproject.toml
[project]
name = "pegasus"
version = "0.1.0"
description = "MontgomeryAI — Civic navigator platform for Montgomery, Alabama"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "brightdata>=0.4.0.4",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.0",
    "langchain>=1.0",
    "langchain-core>=0.3",
    "langchain-google-genai>=4.0",
    "langgraph>=0.3",
    "sse-starlette>=2.0",
    "python-dotenv>=1.0",
    "redis>=7.0.0",
    "uvicorn>=0.41.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.0",
    "PyJWT[crypto]>=2.8",
    "email-validator>=2.0",
    "httpx>=0.28",
    "alembic>=1.18.4",
    "celery[redis]>=5.3",
    "python-multipart>=0.0.22",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "langfuse>=2.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
addopts = "--tb=short"
markers = [
    "unit: pure unit tests with no I/O",
    "integration: tests that cross module boundaries",
    "e2e: end-to-end workflow tests",
    "contract: API/schema contract tests",
    "negative: tests for error paths and invalid inputs",
    "edge: edge case tests (empty, min/max, boundary)",
]

[tool.coverage.run]
source = ["backend"]
